requirements.txt
# JobFinder Dependencies

# Web Framework
Flask==2.3.3
Werkzeug==2.3.7

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# AI Integration
openai>=1.40.0
google-generativeai==0.4.0

# Configuration Management
tomli==2.0.1
toml==0.10.2

# Database (SQLite is included with Python)
# No additional dependencies needed for SQLite

# Utility Libraries
python-dotenv==1.0.0

# Authentication & Security
flask-login==0.6.3
authlib==1.2.1
cryptography==41.0.7

# Development Dependencies (optional)
# Uncomment if needed for development
# pytest==7.4.3
# pytest-flask==1.3.0
# black==23.11.0
# flake8==6.1.0